Removed toggle button on pages other than About, by using a new navbar.
Home needed to have its page refreshed in order for the toggle button to be in the activated position when dark mode is on, however refreshing gave a 404 because of Github Pages. So I rendered a new navbar for each page that is not the About page. The user is now limited in that they can only set dark/light mode in About page, and cannot refresh the page when on subpages.