Edited Navbar and tabs to use React useEffect()
This is a great change because now clicking on subpages will not force refresh the page, which caused a split second white flash that made the web app look like it was rendering slow. Also removed SimpleNavbar thanks to this fix, which was a separate similar navbar used only to take out the toggle button element, because being redirected meant the page wasn't being reloaded (and thus toggle button position was being reset even if dark mode was on).